+++ app/schemas/user_info_schema.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class UserInfoSchema(BaseModel):
+    sub: str
+    username: str
+    email: str
+    email_verified: bool